requirements.txt
paho-mqtt==2.1.0
psutil==7.1.0
python-dotenv==1.1.1
